feat: GraphView의 범위를 넘어가는 경우에만 segmentWidth 조정 #20

+++ GyroData/Utility/GraphView/GraphView.swift
@@ -74,8 +74,8 @@ class GraphView: UIView {
         currentSegment?.add(motions)
     }
     
-    func setupSegmentSize(width: Double = GraphNumber.segmentWidth, height: Double) {
-        segmentWidth = CGFloat(width)
+    func setupSegmentSize(width: CGFloat = GraphNumber.segmentWidth, height: Double) {
+        segmentWidth = width
         segmentHeight = height
     }
